requirements.txt
Flask
psycopg2-binary
requests
python-dotenv
twilio
openai
numpy
pandas
faiss-cpu
sentence-transformers
flask